requirements.txt
Django==1.8.5
django-filter==0.11.0
django-rest-framework-mongoengine==2.0.2
django-rest-swagger==0.3.4
djangorestframework==3.0.5
Markdown==2.6.5
mongoengine==0.9.0
pymongo==2.8.1
PyYAML==3.11
wheel==0.24.0
